Draw directly in damage event
WS_EX_LAYER style means the system is buffering the window contents for us, so
in theory, all we need to do is draw it's contents once in an inital WM_PAINT,
and then update damaged regions when damage events occur, which we can do
synchronously in the damage event callback handler.

We can't actually constrain when the system does a WM_PAINT, or for what area,
so improving xcwm_image_copy_full() so we can ask for an arbitrary rectangle of
the window image would be good.

We could avoid this by using UpdateLayeredWindow(), but that requires us to keep
the full window image around ourselves.