Make window creation and destruction be synchronous
At the moment, because Windows windows must be created and destroyed by the
thread that we want to process events with, XCWM_EVENT_WINDOW_CREATE just posts
a message to another thread to create the window, we can then receive and try to
process additional events (usually damage) on that window before the window
creation has completed.

Likewise XCWM_EVENT_WINDOW_DELETE just posts a message to another thread to
delete the window, then releases the xcwm_window_t.  If a different WM_* message
is recieved for that window before the WM_DELETE is processed, the xcw_window_t
pointer stored in the WIN_XCWM_PROP property will be invalid.

Block the processing of additional X events until window creation or destruction
completes.